Add files via upload
This is a web-based application that uses PHP language with the slim framework. This application includes two roles, Admin and Customer.
Admin can add, edit, and modify the product category and the products and also change the status of the customer.
Customers have to register first then, they can log in to their accounts. Then they can add products to the cart and login into their account and proceed further.